refactor: move the product delete button to a separate file
